requirements.txt
sty==1.0.0b11
tqdm==4.35.0
wfuzz==2.4
requests==2.21.0
python_hosts==0.4.6
psutil==5.6.7
python_libnmap==0.7.0
beautifulsoup4==4.8.0
termcolor==1.1.0
PyYAML==5.1.2
xmltodict
pyasn1
impacket
python-ldap
